Clarifies the responsibility of the BatchContextBuilder, giving it its own file. KGCNs now control the building of context arrays
